+++ src/test/kotlin/it/unibo/protelis/web/execution/simulated/SimulatedProtelisEngineTest.kt
@@ -1,15 +1,14 @@
-/*
 package it.unibo.protelis.web.execution.simulated
 
-import io.vertx.core.logging.Logger
-import io.vertx.core.logging.LoggerFactory
 import it.unibo.protelis.web.execution.ProtelisObserver
 import it.unibo.protelis.web.execution.ProtelisUpdateMessage
 import java.util.concurrent.CompletableFuture
 import kotlinx.io.errors.IOException
-import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 
 @DisplayName("Test simulated Protelis engine with Alchemist")
 class SimulatedProtelisEngineTest {
@@ -20,14 +19,14 @@ class SimulatedProtelisEngineTest {
   fun testCodeInject() {
     val simulation = SimulatedProtelisEngine()
 
-    var init = false
+    val init = CompletableFuture<ProtelisUpdateMessage>()
     val step = CompletableFuture<ProtelisUpdateMessage>()
-    var done = false
+    val done = CompletableFuture<ProtelisUpdateMessage>()
 
     val monitor = object : ProtelisObserver {
       override fun initialized(update: ProtelisUpdateMessage) {
         logger.debug("Init done")
-        init = true
+        init.complete(update)
       }
 
       override fun stepDone(update: ProtelisUpdateMessage) {
@@ -37,25 +36,28 @@ class SimulatedProtelisEngineTest {
 
       override fun finished(update: ProtelisUpdateMessage) {
         logger.debug("End done")
-        done = true
+        done.complete(update)
       }
     }
 
-    simulation
-      .setup(
-        this::class.java.classLoader.getResource("code.pt")
-          ?.readText()
-          ?: throw IOException("Code file not found"),
-        monitor
-      )
-      .thenCompose { simulation.start() }
-      .thenCompose { step }
-      .thenCompose { simulation.stop() }
-      .get()
+    simulation.setup(
+      this::class.java.classLoader.getResource("code.pt")
+        ?.readText()
+        ?: throw IOException("Code file not found"),
+      monitor
+    )
+    simulation.start()
+    val stepMsg: ProtelisUpdateMessage? = step.get()
+    simulation.stop()
+    val doneMsg: ProtelisUpdateMessage? = done.get()
 
-    Assertions.assertTrue(init, "Simulation should have been initialized")
-    Assertions.assertNotNull(step.getNow(null), "Simulation should have run at least one step")
-    Assertions.assertTrue(done, "Simulation should have been terminated")
+    val initMsg: ProtelisUpdateMessage? = init.getNow(null)
+    logger.debug("Initialization message: $initMsg")
+    logger.debug("First step done message: $stepMsg")
+    logger.debug("Finalization message: $doneMsg")
+
+    assertNotNull(initMsg, "Simulation should have been initialized")
+    assertNotNull(stepMsg, "Simulation should have run at least one step")
+    assertNotNull(doneMsg, "Simulation should have been terminated")
   }
 }
-*/
